Add files via upload
I'd like to note that some of the CSS files for these projects are a work in progress. However, I hope that providing the HTML files will still be of assistance for your reference and understanding of the project structure and layout. Please feel free to use the provided HTML files as a foundation for further styling and design improvements.